Authentication and Authorization using APIs

+++ controller/api/v1/post_api.js
@@ -14,19 +14,21 @@ module.exports.index = async function(req,res){
 module.exports.destroy = async function(req,res){
         try{
         let post = await Post.findById(req.params.id)
-        // if(post.user == req.user.id){
-            await post.remove()
+        if(post.user == req.user.id){
+             post.remove()
             await Comment.deleteMany({post : req.params.id})
 
           
             return res.json(200,{
                 message : "Post and associated comments deleted successfully"
             })
-    //     }
-    //     else{
+        }
+        else{
     //         req.flash('error','Can not delete the post!')
-    //         return res.redirect('back')
-    //    }
+            return res.json(401,{
+                message: "You can not delete the post"
+            })
+       }
     }catch(err){
         console.log('!!!!!!!!!!!!!!!!!!!!Error',err)
     // req.flash('error',err)
